Changes to products API (#253)
This PR moves `MenuItem`s into their own directories instead of residing
amongst the products.

Furthermore, this PR removes the `ChangedProductResponse` and replaces
this with the `ProductResponse`. The `ChangedProductResponse` was used
as the return type for products that was added or updated. Instead of
having multiple DTOs which would each have to be updated whenever a
change was made to the general entity, we now only have one.
The previous implementation of `ChangedProductResponse` did not contain
the value `IsPerk` nor the `Id`. This meant that adding a product in
shifty would not be able to show the new item's ID (nor conduct further
changes) without reloading the whole table of products (i.e. a get
request to the /products/all endpoint).

---------

Co-authored-by: Jonas Anker Rasmussen <jonasanker@gmail.com>