Adds manager access to issue vouchers (#271)
Users with role Manager can now (alongside users with role Board) issue
vouchers, which also includes access to the products/all endpoint.
This is required to list products available to issue products on. (This
endpoint is used for issuing vouchers as well as the product manager but
nowhere else. The endpoint for products that is usable per user group
used by the app is a different endpoint)