Login with magic link (#289)
This feature aims at enabling magic link authentication through mails.
The mail will then provide a deeplink for the correct application
(app/shifty) on valid login and use refresh tokens to stay logged in.

This is only for Shifty at this point, since the implementation is
missing for the app frontend. Once implemented in both places, we should
do a soft roll-over of requiring users to update the app, possibly with
a hard requirement at the beginning of the new semester.

---------

Co-authored-by: Andreas Trøstrup <andreastrstrup@gmail.com>