Fix/warnings revived (#310)
This PR cherry-picks the xml comments created in #183, reduces warnings
by marking properties as required, instead of using constructors.

Also removes unused using statements, and other low hanging warnings.

Note, there is still a fair bit of warnings since the tests haven't been
refactored in this PR. I think the best way to address the warnings from
there, would be to migrate them to use the builders, though that seems
out of scope of this PR.

---------

Co-authored-by: Omid Marfavi <21163286+marfavi@users.noreply.github.com>