Revert "feat: Add owner field to support transfer of tickets (#330)" (#336)
This reverts commit 45549d916b454b60c60176494740d829bd96d853.

Why? I realized that there is no reason for the new column on Ticket,
since we already the information in Purchases table, and a Ticket is
always related to the original purchase, so adding a new column is just
duplicated data.
Something something database normal form.
Who would have thought Introduction to Databases would ever be useful?